+++ src/operators/tensor/implementations/tensor_fp8x23wide.cairo
@@ -71,7 +71,7 @@ impl FP8x23WTensor of TensorTrait<FP8x23W> {
         unravel_index(index, *self.shape)
     }
 
-    fn reshape(self: @Tensor<FP8x23W>, target_shape: Span<usize>) -> Tensor<FP8x23W> {
+    fn reshape(self: @Tensor<FP8x23W>, target_shape: Span<i32>) -> Tensor<FP8x23W> {
         reshape(self, target_shape)
     }
